requirements.txt
numpy==1.10.4
Pillow==3.1.0
progressbar33==2.4
scipy==0.16.1
